bug fixed

+++ app/api/purchase/close/route.ts
@@ -36,6 +36,13 @@ export async function PATCH(req: NextRequest) {
       return new Response(JSON.stringify(response));
     }
 
+    if (!purchase.vendorId || purchase.vendorId.length === 0) {
+      response.status = 400;
+      response.message =
+        "No vendor is assigned to this purchase yet! Please assign before closing this purchase order.";
+      return new Response(JSON.stringify(response));
+    }
+
     if (purchase.barcodeRegister.length === 0) {
       await prisma.purchase.delete({
         where: {
